Add getter methods and move BestillingMapper class #deploy-test-dolly-backend
Added getter methods for occupationTypes, workLoadTypes, and workScheduleTypes in ArbeidsplassenCVDTO. These methods ensure that the corresponding lists are initialized when they are null. Also, the BestillingMapper class has been moved from 'strategy' subdirectory to 'mapper' directory for better organization.
